数列求和抽象方法 返回值改为int类型

+++ src/tools/math/series/Series.java
@@ -1,5 +1,5 @@
 package tools.math.series;
 
 public abstract class Series {
-    public abstract double getSum();
+    public abstract int getSum();
 }
